Move check/types files to identify/types
